backend/requirements.txt
azure-functions
requests>=2.31.0
openai>=1.12.0
azure-storage-blob>=12.19.0
